Added tickType stats route

+++ projects/climbingLog/controllers/routeController.js
@@ -3,7 +3,7 @@ const Route = require('../models/routeModel');
 const routeSchema = require('../validators/routeValidator');
 const {findRoute, updateRouteData, populateRoute, populateRoutes} = require('../services/routeServices');
 const validateSchema = require('../middleware/validateSchema');
-const { deleteAreaIfEmpty } = require('../services/areaServices');
+const { deleteAreaIfEmpty, findOrCreateArea } = require('../services/areaServices');
 
 
 exports.getAllRoutes = [
@@ -39,24 +39,34 @@ exports.updateRoute = [
             const route = await findRoute(req.params.id, req.user._id);
 			
 			// Get the id of the existing area
-			const existingAreaId = route.area
+			const existingAreaId = route.area ? route.area : null;
 			
-			// Check if the new area exists in the database
-			const area = await findOrCreateArea(req.body.area, req.user._id);
-
             // Update the route
             const newData = {
                 name: req.body.name,
                 grade: req.body.grade,
                 colour: req.body.colour,
                 user: req.user._id,
-				area: area._id
+				// area: area._id
             };
 
+			if (req.body.area && req.body.area.name !== '') {
+				const area = await findOrCreateArea(req.body.area, req.user._id)
+				newData.area = area._id;
+			}
+
             const updatedRoute = updateRouteData(route, newData);
+
+			// If no area is provided and there is an existing area, remove the area from the route
+			if (!req.body.area && existingAreaId) {
+				updatedRoute.area = null;
+			}
+
             await updatedRoute.save();
 
-			await deleteAreaIfEmpty(req.user._id, existingAreaId)
+			if (existingAreaId) {
+				await deleteAreaIfEmpty(req.user._id, existingAreaId)
+			}
 
 			// Populate the updated route
 			const populatedRoute = await populateRoute(updatedRoute);
